requirements.txt
streamlit==1.44.1
boto3==1.34.131
botocore==1.34.131
python-dotenv==1.0.1